// src/infrastructure/persistence/supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      entity: {
        Row: {
          created_at: string
          display_number: number
          id: string
          name: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          display_number: number
          id?: string
          name: string
          updated_at?: string
        }
        Update: {
          created_at?: string
          display_number?: number
          id?: string
          name?: string
          updated_at?: string
        }
        Relationships: []
      }
      entity_members: {
        Row: {
          created_at: string
          entity_id: string
          role: string
          updated_at: string
          user_id: string
        }
        Insert: {
          created_at?: string
          entity_id: string
          role?: string
          updated_at?: string
          user_id: string
        }
        Update: {
          created_at?: string
          entity_id?: string
          role?: string
          updated_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "entity_members_entity_id_fkey"
            columns: ["entity_id"]
            isOneToOne: false
            referencedRelation: "entity"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}